# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MEGA-TIK Finance & Subscribers ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /areas:
    get:
      operationId: listAreas
      tags: [areas]
      summary: List areas
      responses:
        "200":
          description: List of areas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Area"
    post:
      operationId: createArea
      tags: [areas]
      summary: Create area
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAreaRequest"
      responses:
        "201":
          description: Created area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
  /areas/{id}:
    get:
      operationId: getArea
      tags: [areas]
      summary: Get area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
    put:
      operationId: updateArea
      tags: [areas]
      summary: Update area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAreaRequest"
      responses:
        "200":
          description: Updated area
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Area"
    delete:
      operationId: deleteArea
      tags: [areas]
      summary: Delete area
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List plans
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [plans]
      summary: Create plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Created plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /plans/{id}:
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    put:
      operationId: updatePlan
      tags: [plans]
      summary: Update plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "200":
          description: Updated plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [plans]
      summary: Delete plan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /debts:
    get:
      operationId: listDebts
      tags: [debts]
      summary: List debts
      parameters:
        - name: subscriber_id
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [unpaid, partial, paid]
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Paginated list of debts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtListResponse"
  /debts/{id}:
    get:
      operationId: getDebt
      tags: [debts]
      summary: Get debt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Debt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Debt"
  /debts/{id}/pay:
    post:
      operationId: payDebt
      tags: [debts]
      summary: Pay debt (partial or full)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayDebtRequest"
      responses:
        "200":
          description: Debt payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebtPayment"
...
  /price-tiers:
    get:
      operationId: listPriceTiers
      tags: [price-tiers]
      summary: List price tiers
      responses:
        "200":
          description: List of price tiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PriceTier"
    post:
      operationId: createPriceTier
      tags: [price-tiers]
      summary: Create price tier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceTierRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
  /price-tiers/{id}:
    put:
      operationId: updatePriceTier
      tags: [price-tiers]
      summary: Update price tier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePriceTierRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceTier"
    delete:
      operationId: deletePriceTier
      tags: [price-tiers]
      summary: Delete price tier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, supervisor, accountant, collector]
        locale:
          type: string
          enum: [ar, en]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
...
    Area:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        subscriberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
    CreateAreaRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
      required:
        - nameAr
        - nameEn
    Plan:
      type: object
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: string
        price:
          type: number
        gigabytes:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
        subscriberCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nameAr
        - nameEn
        - price
        - gigabytes
        - durationDays
        - isActive
    CreatePlanRequest:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        price:
          type: number
        gigabytes:
          type: number
        durationDays:
          type: integer
        isActive:
          type: boolean
      required:
        - nameAr
        - nameEn
        - price
        - gigabytes
        - durationDays
...
    Debt:
      type: object
      properties:
        id:
          type: integer
        subscriberId:
          type: integer
        subscriberNameAr:
          type: string
        subscriberNameEn:
          type: string
        renewalId:
          type: integer
        amount:
          type: number
        paidAmount:
          type: number
        remainingAmount:
          type: number
        status:
          type: string
          enum: [unpaid, partial, paid]
        dueDate:
          type: string
          format: date
        notes:
          type: string
        payments:
          type: array
          items:
            $ref: "#/components/schemas/DebtPayment"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - subscriberId
        - amount
        - paidAmount
        - remainingAmount
        - status
    DebtListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Debt"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
        totalDebt:
          type: number
        totalPaid:
          type: number
        totalRemaining:
          type: number
      required:
        - data
        - total
        - page
        - limit
        - totalPages
        - totalDebt
        - totalPaid
        - totalRemaining
    DebtPayment:
      type: object
      properties:
        id:
          type: integer
        debtId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, bank_transfer]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - debtId
        - amount
        - paymentMethod
    PayDebtRequest:
      type: object
      properties:
        amount:
          type: number
        paymentMethod:
          type: string
          enum: [cash, vodafone_cash, bank_transfer]
        notes:
          type: string
      required:
        - amount
        - paymentMethod
...
    PriceTier:
      type: object
      properties:
        id:
          type: integer
        minQuantity:
          type: number
        maxQuantity:
          type: number
        pricePerGb:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - minQuantity
        - maxQuantity
        - pricePerGb
        - isActive
    CreatePriceTierRequest:
      type: object
      properties:
        minQuantity:
          type: number
        maxQuantity:
          type: number
        pricePerGb:
          type: number
        isActive:
          type: boolean
      required:
        - minQuantity
        - maxQuantity
        - pricePerGb
...